remove: delete dmdata_oauth package and its related files, including .gitignore, analysis_options.yaml, build.yaml, pubspec.yaml, and source files. Update pubspec.lock and README.md in dmdata_oauth_flutter to reflect changes.
