+++ .gitignore
@@ -5,3 +5,5 @@
 .dart_tool/
 
 pubspec_overrides.yaml
+
+custom_lint.log
