--- packages/dmdata_oauth/lib/dmdata_oauth.dart
@@ -1,1 +1,0 @@
-library;
