+++ packages/dmdata_oauth_api_client/README.md
@@ -1,0 +1,8 @@
+# dmdata_oauth_api_client
+
+DMDATA APIのOAuth認証を扱うためのDartパッケージです。
+
+## 機能
+
+- dmdata.jpのOAuth 2.0 v1の認証フロー用のAPI Client
+- 認証用のURLを生成する`OauthAuthorizationUrl`
