+++ packages/dmdata_oauth_api_client/lib/dmdata_oauth_api_client.dart
@@ -1,0 +1,3 @@
+library;
+
+export 'src/src.dart';
