+++ packages/dmdata_oauth_api_client/lib/src/model/oauth_error_response.dart
