+++ packages/dmdata_oauth_api_client/lib/src/model/oauth_token_response.dart
