+++ packages/dmdata_oauth_flutter/README.md
@@ -1,39 +1,19 @@
-<!-- 
-This README describes the package. If you publish this package to pub.dev,
-this README's contents appear on the landing page for your package.
+# dmdata_oauth_flutter
 
-For information about how to write a good package README, see the guide for
-[writing package pages](https://dart.dev/tools/pub/writing-package-pages). 
+DMDATAのOAuth2認証をFlutterアプリケーションで簡単に実装するためのパッケージです。
 
-For general information about developing packages, see the Dart guide for
-[creating packages](https://dart.dev/guides/libraries/create-packages)
-and the Flutter guide for
-[developing packages and plugins](https://flutter.dev/to/develop-packages). 
--->
+## 機能
 
-TODO: Put a short description of the package here that helps potential users
-know whether this package might be useful for them.
+- OAuth2.0認可コードフロー
+  - PKCE対応
+  - カスタムスキーム対応
 
-## Features
+## セキュリティに関する注意
 
-TODO: List what your package can do. Maybe include images, gifs, or videos.
+1. リダイレクトURIはアプリケーション固有のスキームを使用してください
+1. PKCEを必ず有効にしてください
+1. stateパラメータは自動的に生成・検証されます
 
-## Getting started
+## ライセンス
 
-TODO: List prerequisites and provide or point to information on how to
-start using the package.
-
-## Usage
-
-TODO: Include short and useful examples for package users. Add longer examples
-to `/example` folder. 
-
-```dart
-const like = 'sample';
-```
-
-## Additional information
-
-TODO: Tell users more about the package: where to find more information, how to 
-contribute to the package, how to file issues, what response they can expect 
-from the package authors, and more.
+MIT License
