--- packages/dmdata_oauth_flutter/lib/dmdata_oauth.dart
@@ -1,8 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library;
-
-export 'src/dmdata_oauth_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
