--- packages/dmdata_oauth_flutter/lib/src/dmdata_oauth_base.dart
@@ -1,6 +1,0 @@
-// TODO: Put public facing types in this file.
-
-/// Checks if you are awesome. Spoiler: you are.
-class Awesome {
-  bool get isAwesome => true;
-}
